requirements.txt
streamlit==1.41.1
litellm>=1.82.0
pandas==2.2.3
numpy==1.26.4
matplotlib==3.10.0
altair==5.5.0
python-dotenv==1.0.1
requests==2.32.3
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
trafilatura>=2.0.0
